+++ libs/ymir-core/include/ymir/media/binary_reader/binary_reader_impl.hpp
@@ -2,6 +2,7 @@
 
 // This header file includes all implementations of IBinaryReader for ease of use.
 
+#include "binary_reader_composite.hpp"
 #include "binary_reader_file.hpp"
 #include "binary_reader_mem.hpp"
 #include "binary_reader_mmap.hpp"