+++ libs/ymir-core/src/ymir/media/loader/loader_chd.cpp
@@ -363,7 +363,9 @@ bool Load(std::filesystem::path chdPath, Disc &disc, bool preloadToRAM) {
                 std::make_unique<SharedSubviewBinaryReader>(binaryReader, subviewOffset, frames * track.sectorSize);
             track.startFrameAddress = frameAddress;
             track.endFrameAddress = frameAddress + frames - 1;
+            track.track01FrameAddress = frameAddress;
             track.interleavedSubchannel = false;
+            track.indices.emplace_back(); // index 00
             auto &index = track.indices.emplace_back();
             index.startFrameAddress = track.startFrameAddress;
             index.endFrameAddress = track.endFrameAddress;