+++ libs/ymir-core/src/ymir/media/loader/loader_img_ccd_sub.cpp
@@ -284,6 +284,7 @@ bool Load(std::filesystem::path ccdPath, Disc &disc, bool preloadToRAM) {
         for (int j = 0; j < session.numTracks; j++) {
             const uint32 trackNum = firstTrackNum + j;
             auto &track = session.tracks[trackNum - 1];
+            track.indices.emplace_back(); // index 00
             auto &index = track.indices.emplace_back();
             auto &point = ccdSession.tocEntries[trackNum];
             if (point.point != trackNum) {
@@ -297,6 +298,7 @@ bool Load(std::filesystem::path ccdPath, Disc &disc, bool preloadToRAM) {
 
             track.startFrameAddress = TimestampToFrameAddress(point.min, point.sec, point.frame);
             index.startFrameAddress = track.startFrameAddress;
+            track.track01FrameAddress = track.startFrameAddress;
 
             // Finish previous track
             if (j > 0) {